Auto assign new member role, set talk channel
Rudimentary code for assigning a default role to new members of a server on join. This is not yet robust.
Also, command for setting the"talkChannel" that the bot will post messages to. At the moment, the variable assignment doesn't stick. Code will need to be refactored once this is functional.